+++ packages/vue2/src/common/index.js
@@ -1,11 +1,9 @@
-import "./global";
-import filters from "./filters";
-import * as directives from "./directives";
-export * from "./mixins";
-
-export * from "./plugins";
-export * from "./router";
+import './global';
+import filters from './filters';
+import * as directives from './directives';
+export * from './mixins';
+export * from './plugins';
 
 export { filters, directives };
 
-export * from "@lcap/basic-template";
+export * from '@lcap/basic-template';
