+++ packages/vue2/src/common/mixins/cache-variable.js
