+++ packages/vue2/src/common/mixins/index.js
@@ -1,1 +1,1 @@
-export * from "./localCacheVariableMixin";
+export * from './cache-variable';
