+++ packages/vue2/src/common/plugins/data-types/index.js
@@ -1,4 +1,4 @@
-import { initDataTypes } from "@lcap/basic-template";
+import { initDataTypes } from '@lcap/basic-template';
 
 export default {
   install(Vue, options) {
