+++ packages/vue2/src/common/plugins/index.js
@@ -1,9 +1,9 @@
-import AuthPlugin from "./auth";
-import DataTypesPlugin from "./dataTypes";
-import LogicsPlugin from "./logic";
-import RouterPlugin from "./router";
-import ServicesPlugin from "./service";
-import UtilsPlugin from "./utils";
-import ProcessPlugin from "./process";
+import AuthPlugin from './auth';
+import DataTypesPlugin from './data-types';
+import LogicsPlugin from './logic';
+import RouterPlugin from './router';
+import ServicesPlugin from './service';
+import UtilsPlugin from './utils';
+import ProcessPlugin from './process';
 
 export { AuthPlugin, DataTypesPlugin, LogicsPlugin, RouterPlugin, ServicesPlugin, UtilsPlugin, ProcessPlugin };
