--- packages/vue2/src/common/router/index.js
@@ -1,1 +1,0 @@
-export * from "./init";
