+++ packages/vue2/src/mixins/index.js
@@ -1,7 +1,7 @@
-import dataSourceUtils from './dataSourceUtils';
-import enhancedCss from './enhancedCss';
-import keyboardEvent from './keyboardEvent';
-import elementStateAndProp from './elementStateAndProp';
+import dataSourceUtils from './datasource-utils';
+import enhancedCss from './enhanced-css';
+import keyboardEvent from './keyboard-event';
+import elementStateAndProp from './element-state-prop';
 
 export default {
   dataSourceUtils,
