+++ packages/vue2/src/router/init.mobile.js
@@ -1,9 +1,20 @@
-export function createRouter({ routes, VueRouter }) {
+import Vue from 'vue';
+import VueRouter from 'vue-router';
+
+export function createRouter(routes) {
+  Vue.use(VueRouter);
+
   const router = new VueRouter({
     mode: 'history',
     base: window.LcapMicro?.routePrefix || process.env.BASE_URL,
     routes,
   });
+
+  /**
+   * 将路由实例挂载到window对象
+   * 作用：提供全局访问路由实例的能力
+   */
+  window.VueRouterInstance = router;
 
   router.afterEach((to, from) => {
     const saveList = [
@@ -23,5 +34,6 @@ export function createRouter({ routes, VueRouter }) {
       }
     }
   });
+
   return router;
 }
