+++ packages/vue2/src/router/init.pc.js
@@ -1,11 +1,22 @@
 /**
  * 创建路由实例
  */
+import Vue from 'vue';
+import VueRouter from 'vue-router';
 
-export function createRouter({ routes, VueRouter }) {
-  return new VueRouter({
+export function createRouter(routes) {
+  Vue.use(VueRouter);
+
+  const router = new VueRouter({
     mode: 'history',
     base: window.LcapMicro?.routePrefix || process.env.BASE_URL,
     routes,
   });
+  /**
+   * 将路由实例挂载到window对象
+   * 作用：提供全局访问路由实例的能力
+   */
+  window.VueRouterInstance = router;
+
+  return router;
 }
