+++ packages/vue2/src/set-config.pc.js
@@ -1,12 +1,14 @@
 import Vue from 'vue';
 import { setConfig } from '@/common';
 
-import { utils } from './plugins/dataTypes';
+import { utils } from './plugins/data-types';
 import { destination, back, go } from './plugins/router';
-import { createRouter } from './router';
 
-// 设置core config
+/**
+ * 给基础库设置当前平台（PC端）相关特性配置
+ */
 setConfig({
+  // 全局toast
   toast: {
     show(msg) {
       if (typeof Vue.prototype?.$toast?.show === 'function') {
@@ -23,12 +25,13 @@ setConfig({
       console.warn('请在Vue.prototype上挂载$toast.error方法');
     },
   },
+  // 路由跳转
   router: {
     destination,
     back,
     go,
-    createRouter,
   },
+  // 工具函数
   utils: {
     ...utils,
     showMessage(msg) {
@@ -39,6 +42,10 @@ setConfig({
       console.warn('请在Vue.prototype上挂载$toast.show方法');
     },
   },
+  /**
+   * 构造响应式对象
+   * 主要给全局变量提供响应式支持
+   */
   reactive: (obj) => {
     return new Vue({
       data: {
@@ -46,6 +53,9 @@ setConfig({
       },
     });
   },
+  /**
+   * 全局属性设置
+   */
   globalProperties: {
     set(key, value) {
       window[key] = value;
@@ -55,39 +65,27 @@ setConfig({
       return Vue.prototype[key];
     },
   },
-  configureRequest(_options, axios) {
-    axios.interceptors.response.use(
-      function onSuccess(response) {
-        if (response.headers.authorization) {
-          response.data.authorization = response.headers.authorization;
-        }
-        return response;
-      },
-      function onError(error) {
-        return Promise.reject(error);
-      },
-    );
-
+  /**
+   * 自定义配置axios
+   * 可覆盖基础库内置的axios配置
+   */
+  configureRequest(_options, _axios) {
     /**
      * options配置参考
      * https://axios-http.com/zh/docs/req_config
      */
-
     // 修改请求baseURL
     // _options.baseURL = 'https://some-domain.com/api';
-
     // 增加额外的请求头
     // _options.headers = {
     //     ...(_options.headers || {}),
     //     key1: 'value1',
     // }
-
     // 增加额外的请求参数（带在请求链接上）
     // _options.params = {
     //     ...(_options.params || {}),
     //     key2: 'value2',
     // };
-
     // 增加额外的请求参数（带在请求体上）
     // _options.data = {
     //     ...(_options.data || {}),
